requirements.txt
pyqt5
qtpy
opencv-python
scipy
paddleseg
albumentations>0.5
cython
pyyaml
numpy
scikit-image
opencv-python-headless
Pillow
matplotlib
imgaug>=0.4
graphviz
tqdm
easydict
future
cffi
ninja